+++ terraform/product/outputs.tf
@@ -1,18 +1,32 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-output "app_name" {
+output "ingress_requires" {
+  value = {
+    certificates = "certificates"
+    logging      = "logging"
+  }
+}
+
+output "ingress_provides" {
+  value = {
+    grafana_dashboard = "grafana-dashboard"
+    metrics_endpoint  = "metrics-endpoint"
+  }
+}
+
+output "netbox_app_name" {
   description = "Name of the deployed application."
   value       = module.netbox_k8s.app_name
 }
 
-output "requires" {
+output "netbox_requires" {
   value = {
     logging = "logging"
   }
 }
 
-output "provides" {
+output "netbox_provides" {
   value = {
     grafana_dashboard = "grafana-dashboard"
     metrics_endpoint  = "metrics-endpoint"