+++ docs/changelog.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Place any unreleased changes here, that are subject to release in coming versions :).
 
+### 2025-11-11
+
+- Add missing outputs to the Terraform product module.
+
 ## 2025-09-25
 
 * Update documentation.